Accessing FRED This function can pull time series from the FRED database: https://fred.stlouisfed.org

Accessing FRED This function can pull time series from the FRED database: https://fred.stlouisfed.org

get_econ_series(series_id,
series_name = NULL,observation_start = NULL,
 observation_end = NULL)

Arguments

series_id
FRED times series ID, the unique id retrieved from fed database.

series_name
Choose a name for the series column in output. Default: series_id.

observation_start
Choose a name for the series column in output. Default: series_id.

observation_end
Date of last observation in "yyyy-mm-dd" format. Default: Last observation available.

Examples

get_econ_series("UNRATE", "unemploymentRate")

#>           date unemploymentRate
#> 1   1948-01-01              3.4
#> 2   1948-02-01              3.8
#> 3   1948-03-01              4.0
#> 4   1948-04-01              3.9
#> 5   1948-05-01              3.5
#> 6   1948-06-01              3.6
#> 7   1948-07-01              3.6
#> 8   1948-08-01              3.9
#> 9   1948-09-01              3.8
#> 10  1948-10-01              3.7
#> 11  1948-11-01              3.8
#> 12  1948-12-01              4.0
#> 13  1949-01-01              4.3
#> 14  1949-02-01              4.7
#> 15  1949-03-01              5.0
#> 16  1949-04-01              5.3
#> 17  1949-05-01              6.1
#> 18  1949-06-01              6.2
#> 19  1949-07-01              6.7
#> 20  1949-08-01              6.8
#> 21  1949-09-01              6.6
#> 22  1949-10-01              7.9
#> 23  1949-11-01              6.4
#> 24  1949-12-01              6.6
#> 25  1950-01-01              6.5
#> 26  1950-02-01              6.4
#> 27  1950-03-01              6.3
#> 28  1950-04-01              5.8
#> 29  1950-05-01              5.5
#> 30  1950-06-01              5.4
#> 31  1950-07-01              5.0
#> 32  1950-08-01              4.5
#> 33  1950-09-01              4.4
#> 34  1950-10-01              4.2
#> 35  1950-11-01              4.2
#> 36  1950-12-01              4.3
#> 37  1951-01-01              3.7
#> 38  1951-02-01              3.4
#> 39  1951-03-01              3.4
#> 40  1951-04-01              3.1
#> 41  1951-05-01              3.0
#> 42  1951-06-01              3.2
#> 43  1951-07-01              3.1
#> 44  1951-08-01              3.1
#> 45  1951-09-01              3.3
#> 46  1951-10-01              3.5
#> 47  1951-11-01              3.5
#> 48  1951-12-01              3.1
#> 49  1952-01-01              3.2
#> 50  1952-02-01              3.1
#> 51  1952-03-01              2.9
#> 52  1952-04-01              2.9
#> 53  1952-05-01              3.0
#> 54  1952-06-01              3.0
#> 55  1952-07-01              3.2
#> 56  1952-08-01              3.4
#> 57  1952-09-01              3.1
#> 58  1952-10-01              3.0
#> 59  1952-11-01              2.8
#> 60  1952-12-01              2.7
#> 61  1953-01-01              2.9
#> 62  1953-02-01              2.6
#> 63  1953-03-01              2.6
#> 64  1953-04-01              2.7
#> 65  1953-05-01              2.5
#> 66  1953-06-01              2.5
#> 67  1953-07-01              2.6
#> 68  1953-08-01              2.7
#> 69  1953-09-01              2.9
#> 70  1953-10-01              3.1
#> 71  1953-11-01              3.5
#> 72  1953-12-01              4.5
#> 73  1954-01-01              4.9
#> 74  1954-02-01              5.2
#> 75  1954-03-01              5.7
#> 76  1954-04-01              5.9
#> 77  1954-05-01              5.9
#> 78  1954-06-01              5.6
#> 79  1954-07-01              5.8
#> 80  1954-08-01              6.0
#> 81  1954-09-01              6.1
#> 82  1954-10-01              5.7
#> 83  1954-11-01              5.3
#> 84  1954-12-01              5.0
#> 85  1955-01-01              4.9
#> 86  1955-02-01              4.7
#> 87  1955-03-01              4.6
#> 88  1955-04-01              4.7
#> 89  1955-05-01              4.3
#> 90  1955-06-01              4.2
#> 91  1955-07-01              4.0
#> 92  1955-08-01              4.2
#> 93  1955-09-01              4.1
#> 94  1955-10-01              4.3
#> 95  1955-11-01              4.2
#> 96  1955-12-01              4.2
#> 97  1956-01-01              4.0
#> 98  1956-02-01              3.9
#> 99  1956-03-01              4.2
#> 100 1956-04-01              4.0
#> 101 1956-05-01              4.3
#> 102 1956-06-01              4.3
#> 103 1956-07-01              4.4
#> 104 1956-08-01              4.1
#> 105 1956-09-01              3.9
#> 106 1956-10-01              3.9
#> 107 1956-11-01              4.3
#> 108 1956-12-01              4.2
#> 109 1957-01-01              4.2
#> 110 1957-02-01              3.9
#> 111 1957-03-01              3.7
#> 112 1957-04-01              3.9
#> 113 1957-05-01              4.1
#> 114 1957-06-01              4.3
#> 115 1957-07-01              4.2
#> 116 1957-08-01              4.1
#> 117 1957-09-01              4.4
#> 118 1957-10-01              4.5
#> 119 1957-11-01              5.1
#> 120 1957-12-01              5.2
#> 121 1958-01-01              5.8
#> 122 1958-02-01              6.4
#> 123 1958-03-01              6.7
#> 124 1958-04-01              7.4
#> 125 1958-05-01              7.4
#> 126 1958-06-01              7.3
#> 127 1958-07-01              7.5
#> 128 1958-08-01              7.4
#> 129 1958-09-01              7.1
#> 130 1958-10-01              6.7
#> 131 1958-11-01              6.2
#> 132 1958-12-01              6.2
#> 133 1959-01-01              6.0
#> 134 1959-02-01              5.9
#> 135 1959-03-01              5.6
#> 136 1959-04-01              5.2
#> 137 1959-05-01              5.1
#> 138 1959-06-01              5.0
#> 139 1959-07-01              5.1
#> 140 1959-08-01              5.2
#> 141 1959-09-01              5.5
#> 142 1959-10-01              5.7
#> 143 1959-11-01              5.8
#> 144 1959-12-01              5.3
#> 145 1960-01-01              5.2
#> 146 1960-02-01              4.8
#> 147 1960-03-01              5.4
#> 148 1960-04-01              5.2
#> 149 1960-05-01              5.1
#> 150 1960-06-01              5.4
#> 151 1960-07-01              5.5
#> 152 1960-08-01              5.6
#> 153 1960-09-01              5.5
#> 154 1960-10-01              6.1
#> 155 1960-11-01              6.1
#> 156 1960-12-01              6.6
#> 157 1961-01-01              6.6
#> 158 1961-02-01              6.9
#> 159 1961-03-01              6.9
#> 160 1961-04-01              7.0
#> 161 1961-05-01              7.1
#> 162 1961-06-01              6.9
#> 163 1961-07-01              7.0
#> 164 1961-08-01              6.6
#> 165 1961-09-01              6.7
#> 166 1961-10-01              6.5
#> 167 1961-11-01              6.1
#> 168 1961-12-01              6.0
#> 169 1962-01-01              5.8
#> 170 1962-02-01              5.5
#> 171 1962-03-01              5.6
#> 172 1962-04-01              5.6
#> 173 1962-05-01              5.5
#> 174 1962-06-01              5.5
#> 175 1962-07-01              5.4
#> 176 1962-08-01              5.7
#> 177 1962-09-01              5.6
#> 178 1962-10-01              5.4
#> 179 1962-11-01              5.7
#> 180 1962-12-01              5.5
#> 181 1963-01-01              5.7
#> 182 1963-02-01              5.9
#> 183 1963-03-01              5.7
#> 184 1963-04-01              5.7
#> 185 1963-05-01              5.9
#> 186 1963-06-01              5.6
#> 187 1963-07-01              5.6
#> 188 1963-08-01              5.4
#> 189 1963-09-01              5.5
#> 190 1963-10-01              5.5
#> 191 1963-11-01              5.7
#> 192 1963-12-01              5.5
#> 193 1964-01-01              5.6
#> 194 1964-02-01              5.4
#> 195 1964-03-01              5.4
#> 196 1964-04-01              5.3
#> 197 1964-05-01              5.1
#> 198 1964-06-01              5.2
#> 199 1964-07-01              4.9
#> 200 1964-08-01              5.0
#> 201 1964-09-01              5.1
#> 202 1964-10-01              5.1
#> 203 1964-11-01              4.8
#> 204 1964-12-01              5.0
#> 205 1965-01-01              4.9
#> 206 1965-02-01              5.1
#> 207 1965-03-01              4.7
#> 208 1965-04-01              4.8
#> 209 1965-05-01              4.6
#> 210 1965-06-01              4.6
#> 211 1965-07-01              4.4
#> 212 1965-08-01              4.4
#> 213 1965-09-01              4.3
#> 214 1965-10-01              4.2
#> 215 1965-11-01              4.1
#> 216 1965-12-01              4.0
#> 217 1966-01-01              4.0
#> 218 1966-02-01              3.8
#> 219 1966-03-01              3.8
#> 220 1966-04-01              3.8
#> 221 1966-05-01              3.9
#> 222 1966-06-01              3.8
#> 223 1966-07-01              3.8
#> 224 1966-08-01              3.8
#> 225 1966-09-01              3.7
#> 226 1966-10-01              3.7
#> 227 1966-11-01              3.6
#> 228 1966-12-01              3.8
#> 229 1967-01-01              3.9
#> 230 1967-02-01              3.8
#> 231 1967-03-01              3.8
#> 232 1967-04-01              3.8
#> 233 1967-05-01              3.8
#> 234 1967-06-01              3.9
#> 235 1967-07-01              3.8
#> 236 1967-08-01              3.8
#> 237 1967-09-01              3.8
#> 238 1967-10-01              4.0
#> 239 1967-11-01              3.9
#> 240 1967-12-01              3.8
#> 241 1968-01-01              3.7
#> 242 1968-02-01              3.8
#> 243 1968-03-01              3.7
#> 244 1968-04-01              3.5
#> 245 1968-05-01              3.5
#> 246 1968-06-01              3.7
#> 247 1968-07-01              3.7
#> 248 1968-08-01              3.5
#> 249 1968-09-01              3.4
#> 250 1968-10-01              3.4
#> 251 1968-11-01              3.4
#> 252 1968-12-01              3.4
#> 253 1969-01-01              3.4
#> 254 1969-02-01              3.4
#> 255 1969-03-01              3.4
#> 256 1969-04-01              3.4
#> 257 1969-05-01              3.4
#> 258 1969-06-01              3.5
#> 259 1969-07-01              3.5
#> 260 1969-08-01              3.5
#> 261 1969-09-01              3.7
#> 262 1969-10-01              3.7
#> 263 1969-11-01              3.5
#> 264 1969-12-01              3.5
#> 265 1970-01-01              3.9
#> 266 1970-02-01              4.2
#> 267 1970-03-01              4.4
#> 268 1970-04-01              4.6
#> 269 1970-05-01              4.8
#> 270 1970-06-01              4.9
#> 271 1970-07-01              5.0
#> 272 1970-08-01              5.1
#> 273 1970-09-01              5.4
#> 274 1970-10-01              5.5
#> 275 1970-11-01              5.9
#> 276 1970-12-01              6.1
#> 277 1971-01-01              5.9
#> 278 1971-02-01              5.9
#> 279 1971-03-01              6.0
#> 280 1971-04-01              5.9
#> 281 1971-05-01              5.9
#> 282 1971-06-01              5.9
#> 283 1971-07-01              6.0
#> 284 1971-08-01              6.1
#> 285 1971-09-01              6.0
#> 286 1971-10-01              5.8
#> 287 1971-11-01              6.0
#> 288 1971-12-01              6.0
#> 289 1972-01-01              5.8
#> 290 1972-02-01              5.7
#> 291 1972-03-01              5.8
#> 292 1972-04-01              5.7
#> 293 1972-05-01              5.7
#> 294 1972-06-01              5.7
#> 295 1972-07-01              5.6
#> 296 1972-08-01              5.6
#> 297 1972-09-01              5.5
#> 298 1972-10-01              5.6
#> 299 1972-11-01              5.3
#> 300 1972-12-01              5.2
#> 301 1973-01-01              4.9
#> 302 1973-02-01              5.0
#> 303 1973-03-01              4.9
#> 304 1973-04-01              5.0
#> 305 1973-05-01              4.9
#> 306 1973-06-01              4.9
#> 307 1973-07-01              4.8
#> 308 1973-08-01              4.8
#> 309 1973-09-01              4.8
#> 310 1973-10-01              4.6
#> 311 1973-11-01              4.8
#> 312 1973-12-01              4.9
#> 313 1974-01-01              5.1
#> 314 1974-02-01              5.2
#> 315 1974-03-01              5.1
#> 316 1974-04-01              5.1
#> 317 1974-05-01              5.1
#> 318 1974-06-01              5.4
#> 319 1974-07-01              5.5
#> 320 1974-08-01              5.5
#> 321 1974-09-01              5.9
#> 322 1974-10-01              6.0
#> 323 1974-11-01              6.6
#> 324 1974-12-01              7.2
#> 325 1975-01-01              8.1
#> 326 1975-02-01              8.1
#> 327 1975-03-01              8.6
#> 328 1975-04-01              8.8
#> 329 1975-05-01              9.0
#> 330 1975-06-01              8.8
#> 331 1975-07-01              8.6
#> 332 1975-08-01              8.4
#> 333 1975-09-01              8.4
#> 334 1975-10-01              8.4
#> 335 1975-11-01              8.3
#> 336 1975-12-01              8.2
#> 337 1976-01-01              7.9
#> 338 1976-02-01              7.7
#> 339 1976-03-01              7.6
#> 340 1976-04-01              7.7
#> 341 1976-05-01              7.4
#> 342 1976-06-01              7.6
#> 343 1976-07-01              7.8
#> 344 1976-08-01              7.8
#> 345 1976-09-01              7.6
#> 346 1976-10-01              7.7
#> 347 1976-11-01              7.8
#> 348 1976-12-01              7.8
#> 349 1977-01-01              7.5
#> 350 1977-02-01              7.6
#> 351 1977-03-01              7.4
#> 352 1977-04-01              7.2
#> 353 1977-05-01              7.0
#> 354 1977-06-01              7.2
#> 355 1977-07-01              6.9
#> 356 1977-08-01              7.0
#> 357 1977-09-01              6.8
#> 358 1977-10-01              6.8
#> 359 1977-11-01              6.8
#> 360 1977-12-01              6.4
#> 361 1978-01-01              6.4
#> 362 1978-02-01              6.3
#> 363 1978-03-01              6.3
#> 364 1978-04-01              6.1
#> 365 1978-05-01              6.0
#> 366 1978-06-01              5.9
#> 367 1978-07-01              6.2
#> 368 1978-08-01              5.9
#> 369 1978-09-01              6.0
#> 370 1978-10-01              5.8
#> 371 1978-11-01              5.9
#> 372 1978-12-01              6.0
#> 373 1979-01-01              5.9
#> 374 1979-02-01              5.9
#> 375 1979-03-01              5.8
#> 376 1979-04-01              5.8
#> 377 1979-05-01              5.6
#> 378 1979-06-01              5.7
#> 379 1979-07-01              5.7
#> 380 1979-08-01              6.0
#> 381 1979-09-01              5.9
#> 382 1979-10-01              6.0
#> 383 1979-11-01              5.9
#> 384 1979-12-01              6.0
#> 385 1980-01-01              6.3
#> 386 1980-02-01              6.3
#> 387 1980-03-01              6.3
#> 388 1980-04-01              6.9
#> 389 1980-05-01              7.5
#> 390 1980-06-01              7.6
#> 391 1980-07-01              7.8
#> 392 1980-08-01              7.7
#> 393 1980-09-01              7.5
#> 394 1980-10-01              7.5
#> 395 1980-11-01              7.5
#> 396 1980-12-01              7.2
#> 397 1981-01-01              7.5
#> 398 1981-02-01              7.4
#> 399 1981-03-01              7.4
#> 400 1981-04-01              7.2
#> 401 1981-05-01              7.5
#> 402 1981-06-01              7.5
#> 403 1981-07-01              7.2
#> 404 1981-08-01              7.4
#> 405 1981-09-01              7.6
#> 406 1981-10-01              7.9
#> 407 1981-11-01              8.3
#> 408 1981-12-01              8.5
#> 409 1982-01-01              8.6
#> 410 1982-02-01              8.9
#> 411 1982-03-01              9.0
#> 412 1982-04-01              9.3
#> 413 1982-05-01              9.4
#> 414 1982-06-01              9.6
#> 415 1982-07-01              9.8
#> 416 1982-08-01              9.8
#> 417 1982-09-01             10.1
#> 418 1982-10-01             10.4
#> 419 1982-11-01             10.8
#> 420 1982-12-01             10.8
#> 421 1983-01-01             10.4
#> 422 1983-02-01             10.4
#> 423 1983-03-01             10.3
#> 424 1983-04-01             10.2
#> 425 1983-05-01             10.1
#> 426 1983-06-01             10.1
#> 427 1983-07-01              9.4
#> 428 1983-08-01              9.5
#> 429 1983-09-01              9.2
#> 430 1983-10-01              8.8
#> 431 1983-11-01              8.5
#> 432 1983-12-01              8.3
#> 433 1984-01-01              8.0
#> 434 1984-02-01              7.8
#> 435 1984-03-01              7.8
#> 436 1984-04-01              7.7
#> 437 1984-05-01              7.4
#> 438 1984-06-01              7.2
#> 439 1984-07-01              7.5
#> 440 1984-08-01              7.5
#> 441 1984-09-01              7.3
#> 442 1984-10-01              7.4
#> 443 1984-11-01              7.2
#> 444 1984-12-01              7.3
#> 445 1985-01-01              7.3
#> 446 1985-02-01              7.2
#> 447 1985-03-01              7.2
#> 448 1985-04-01              7.3
#> 449 1985-05-01              7.2
#> 450 1985-06-01              7.4
#> 451 1985-07-01              7.4
#> 452 1985-08-01              7.1
#> 453 1985-09-01              7.1
#> 454 1985-10-01              7.1
#> 455 1985-11-01              7.0
#> 456 1985-12-01              7.0
#> 457 1986-01-01              6.7
#> 458 1986-02-01              7.2
#> 459 1986-03-01              7.2
#> 460 1986-04-01              7.1
#> 461 1986-05-01              7.2
#> 462 1986-06-01              7.2
#> 463 1986-07-01              7.0
#> 464 1986-08-01              6.9
#> 465 1986-09-01              7.0
#> 466 1986-10-01              7.0
#> 467 1986-11-01              6.9
#> 468 1986-12-01              6.6
#> 469 1987-01-01              6.6
#> 470 1987-02-01              6.6
#> 471 1987-03-01              6.6
#> 472 1987-04-01              6.3
#> 473 1987-05-01              6.3
#> 474 1987-06-01              6.2
#> 475 1987-07-01              6.1
#> 476 1987-08-01              6.0
#> 477 1987-09-01              5.9
#> 478 1987-10-01              6.0
#> 479 1987-11-01              5.8
#> 480 1987-12-01              5.7
#> 481 1988-01-01              5.7
#> 482 1988-02-01              5.7
#> 483 1988-03-01              5.7
#> 484 1988-04-01              5.4
#> 485 1988-05-01              5.6
#> 486 1988-06-01              5.4
#> 487 1988-07-01              5.4
#> 488 1988-08-01              5.6
#> 489 1988-09-01              5.4
#> 490 1988-10-01              5.4
#> 491 1988-11-01              5.3
#> 492 1988-12-01              5.3
#> 493 1989-01-01              5.4
#> 494 1989-02-01              5.2
#> 495 1989-03-01              5.0
#> 496 1989-04-01              5.2
#> 497 1989-05-01              5.2
#> 498 1989-06-01              5.3
#> 499 1989-07-01              5.2
#> 500 1989-08-01              5.2
#> 501 1989-09-01              5.3
#> 502 1989-10-01              5.3
#> 503 1989-11-01              5.4
#> 504 1989-12-01              5.4
#> 505 1990-01-01              5.4
#> 506 1990-02-01              5.3
#> 507 1990-03-01              5.2
#> 508 1990-04-01              5.4
#> 509 1990-05-01              5.4
#> 510 1990-06-01              5.2
#> 511 1990-07-01              5.5
#> 512 1990-08-01              5.7
#> 513 1990-09-01              5.9
#> 514 1990-10-01              5.9
#> 515 1990-11-01              6.2
#> 516 1990-12-01              6.3
#> 517 1991-01-01              6.4
#> 518 1991-02-01              6.6
#> 519 1991-03-01              6.8
#> 520 1991-04-01              6.7
#> 521 1991-05-01              6.9
#> 522 1991-06-01              6.9
#> 523 1991-07-01              6.8
#> 524 1991-08-01              6.9
#> 525 1991-09-01              6.9
#> 526 1991-10-01              7.0
#> 527 1991-11-01              7.0
#> 528 1991-12-01              7.3
#> 529 1992-01-01              7.3
#> 530 1992-02-01              7.4
#> 531 1992-03-01              7.4
#> 532 1992-04-01              7.4
#> 533 1992-05-01              7.6
#> 534 1992-06-01              7.8
#> 535 1992-07-01              7.7
#> 536 1992-08-01              7.6
#> 537 1992-09-01              7.6
#> 538 1992-10-01              7.3
#> 539 1992-11-01              7.4
#> 540 1992-12-01              7.4
#> 541 1993-01-01              7.3
#> 542 1993-02-01              7.1
#> 543 1993-03-01              7.0
#> 544 1993-04-01              7.1
#> 545 1993-05-01              7.1
#> 546 1993-06-01              7.0
#> 547 1993-07-01              6.9
#> 548 1993-08-01              6.8
#> 549 1993-09-01              6.7
#> 550 1993-10-01              6.8
#> 551 1993-11-01              6.6
#> 552 1993-12-01              6.5
#> 553 1994-01-01              6.6
#> 554 1994-02-01              6.6
#> 555 1994-03-01              6.5
#> 556 1994-04-01              6.4
#> 557 1994-05-01              6.1
#> 558 1994-06-01              6.1
#> 559 1994-07-01              6.1
#> 560 1994-08-01              6.0
#> 561 1994-09-01              5.9
#> 562 1994-10-01              5.8
#> 563 1994-11-01              5.6
#> 564 1994-12-01              5.5
#> 565 1995-01-01              5.6
#> 566 1995-02-01              5.4
#> 567 1995-03-01              5.4
#> 568 1995-04-01              5.8
#> 569 1995-05-01              5.6
#> 570 1995-06-01              5.6
#> 571 1995-07-01              5.7
#> 572 1995-08-01              5.7
#> 573 1995-09-01              5.6
#> 574 1995-10-01              5.5
#> 575 1995-11-01              5.6
#> 576 1995-12-01              5.6
#> 577 1996-01-01              5.6
#> 578 1996-02-01              5.5
#> 579 1996-03-01              5.5
#> 580 1996-04-01              5.6
#> 581 1996-05-01              5.6
#> 582 1996-06-01              5.3
#> 583 1996-07-01              5.5
#> 584 1996-08-01              5.1
#> 585 1996-09-01              5.2
#> 586 1996-10-01              5.2
#> 587 1996-11-01              5.4
#> 588 1996-12-01              5.4
#> 589 1997-01-01              5.3
#> 590 1997-02-01              5.2
#> 591 1997-03-01              5.2
#> 592 1997-04-01              5.1
#> 593 1997-05-01              4.9
#> 594 1997-06-01              5.0
#> 595 1997-07-01              4.9
#> 596 1997-08-01              4.8
#> 597 1997-09-01              4.9
#> 598 1997-10-01              4.7
#> 599 1997-11-01              4.6
#> 600 1997-12-01              4.7
#> 601 1998-01-01              4.6
#> 602 1998-02-01              4.6
#> 603 1998-03-01              4.7
#> 604 1998-04-01              4.3
#> 605 1998-05-01              4.4
#> 606 1998-06-01              4.5
#> 607 1998-07-01              4.5
#> 608 1998-08-01              4.5
#> 609 1998-09-01              4.6
#> 610 1998-10-01              4.5
#> 611 1998-11-01              4.4
#> 612 1998-12-01              4.4
#> 613 1999-01-01              4.3
#> 614 1999-02-01              4.4
#> 615 1999-03-01              4.2
#> 616 1999-04-01              4.3
#> 617 1999-05-01              4.2
#> 618 1999-06-01              4.3
#> 619 1999-07-01              4.3
#> 620 1999-08-01              4.2
#> 621 1999-09-01              4.2
#> 622 1999-10-01              4.1
#> 623 1999-11-01              4.1
#> 624 1999-12-01              4.0
#> 625 2000-01-01              4.0
#> 626 2000-02-01              4.1
#> 627 2000-03-01              4.0
#> 628 2000-04-01              3.8
#> 629 2000-05-01              4.0
#> 630 2000-06-01              4.0
#> 631 2000-07-01              4.0
#> 632 2000-08-01              4.1
#> 633 2000-09-01              3.9
#> 634 2000-10-01              3.9
#> 635 2000-11-01              3.9
#> 636 2000-12-01              3.9
#> 637 2001-01-01              4.2
#> 638 2001-02-01              4.2
#> 639 2001-03-01              4.3
#> 640 2001-04-01              4.4
#> 641 2001-05-01              4.3
#> 642 2001-06-01              4.5
#> 643 2001-07-01              4.6
#> 644 2001-08-01              4.9
#> 645 2001-09-01              5.0
#> 646 2001-10-01              5.3
#> 647 2001-11-01              5.5
#> 648 2001-12-01              5.7
#> 649 2002-01-01              5.7
#> 650 2002-02-01              5.7
#> 651 2002-03-01              5.7
#> 652 2002-04-01              5.9
#> 653 2002-05-01              5.8
#> 654 2002-06-01              5.8
#> 655 2002-07-01              5.8
#> 656 2002-08-01              5.7
#> 657 2002-09-01              5.7
#> 658 2002-10-01              5.7
#> 659 2002-11-01              5.9
#> 660 2002-12-01              6.0
#> 661 2003-01-01              5.8
#> 662 2003-02-01              5.9
#> 663 2003-03-01              5.9
#> 664 2003-04-01              6.0
#> 665 2003-05-01              6.1
#> 666 2003-06-01              6.3
#> 667 2003-07-01              6.2
#> 668 2003-08-01              6.1
#> 669 2003-09-01              6.1
#> 670 2003-10-01              6.0
#> 671 2003-11-01              5.8
#> 672 2003-12-01              5.7
#> 673 2004-01-01              5.7
#> 674 2004-02-01              5.6
#> 675 2004-03-01              5.8
#> 676 2004-04-01              5.6
#> 677 2004-05-01              5.6
#> 678 2004-06-01              5.6
#> 679 2004-07-01              5.5
#> 680 2004-08-01              5.4
#> 681 2004-09-01              5.4
#> 682 2004-10-01              5.5
#> 683 2004-11-01              5.4
#> 684 2004-12-01              5.4
#> 685 2005-01-01              5.3
#> 686 2005-02-01              5.4
#> 687 2005-03-01              5.2
#> 688 2005-04-01              5.2
#> 689 2005-05-01              5.1
#> 690 2005-06-01              5.0
#> 691 2005-07-01              5.0
#> 692 2005-08-01              4.9
#> 693 2005-09-01              5.0
#> 694 2005-10-01              5.0
#> 695 2005-11-01              5.0
#> 696 2005-12-01              4.9
#> 697 2006-01-01              4.7
#> 698 2006-02-01              4.8
#> 699 2006-03-01              4.7
#> 700 2006-04-01              4.7
#> 701 2006-05-01              4.6
#> 702 2006-06-01              4.6
#> 703 2006-07-01              4.7
#> 704 2006-08-01              4.7
#> 705 2006-09-01              4.5
#> 706 2006-10-01              4.4
#> 707 2006-11-01              4.5
#> 708 2006-12-01              4.4
#> 709 2007-01-01              4.6
#> 710 2007-02-01              4.5
#> 711 2007-03-01              4.4
#> 712 2007-04-01              4.5
#> 713 2007-05-01              4.4
#> 714 2007-06-01              4.6
#> 715 2007-07-01              4.7
#> 716 2007-08-01              4.6
#> 717 2007-09-01              4.7
#> 718 2007-10-01              4.7
#> 719 2007-11-01              4.7
#> 720 2007-12-01              5.0
#> 721 2008-01-01              5.0
#> 722 2008-02-01              4.9
#> 723 2008-03-01              5.1
#> 724 2008-04-01              5.0
#> 725 2008-05-01              5.4
#> 726 2008-06-01              5.6
#> 727 2008-07-01              5.8
#> 728 2008-08-01              6.1
#> 729 2008-09-01              6.1
#> 730 2008-10-01              6.5
#> 731 2008-11-01              6.8
#> 732 2008-12-01              7.3
#> 733 2009-01-01              7.8
#> 734 2009-02-01              8.3
#> 735 2009-03-01              8.7
#> 736 2009-04-01              9.0
#> 737 2009-05-01              9.4
#> 738 2009-06-01              9.5
#> 739 2009-07-01              9.5
#> 740 2009-08-01              9.6
#> 741 2009-09-01              9.8
#> 742 2009-10-01             10.0
#> 743 2009-11-01              9.9
#> 744 2009-12-01              9.9
#> 745 2010-01-01              9.8
#> 746 2010-02-01              9.8
#> 747 2010-03-01              9.9
#> 748 2010-04-01              9.9
#> 749 2010-05-01              9.6
#> 750 2010-06-01              9.4
#> 751 2010-07-01              9.4
#> 752 2010-08-01              9.5
#> 753 2010-09-01              9.5
#> 754 2010-10-01              9.4
#> 755 2010-11-01              9.8
#> 756 2010-12-01              9.3
#> 757 2011-01-01              9.1
#> 758 2011-02-01              9.0
#> 759 2011-03-01              9.0
#> 760 2011-04-01              9.1
#> 761 2011-05-01              9.0
#> 762 2011-06-01              9.1
#> 763 2011-07-01              9.0
#> 764 2011-08-01              9.0
#> 765 2011-09-01              9.0
#> 766 2011-10-01              8.8
#> 767 2011-11-01              8.6
#> 768 2011-12-01              8.5
#> 769 2012-01-01              8.3
#> 770 2012-02-01              8.3
#> 771 2012-03-01              8.2
#> 772 2012-04-01              8.2
#> 773 2012-05-01              8.2
#> 774 2012-06-01              8.2
#> 775 2012-07-01              8.2
#> 776 2012-08-01              8.1
#> 777 2012-09-01              7.8
#> 778 2012-10-01              7.8
#> 779 2012-11-01              7.7
#> 780 2012-12-01              7.9
#> 781 2013-01-01              8.0
#> 782 2013-02-01              7.7
#> 783 2013-03-01              7.5
#> 784 2013-04-01              7.6
#> 785 2013-05-01              7.5
#> 786 2013-06-01              7.5
#> 787 2013-07-01              7.3
#> 788 2013-08-01              7.2
#> 789 2013-09-01              7.2
#> 790 2013-10-01              7.2
#> 791 2013-11-01              6.9
#> 792 2013-12-01              6.7
#> 793 2014-01-01              6.6
#> 794 2014-02-01              6.7
#> 795 2014-03-01              6.7
#> 796 2014-04-01              6.2
#> 797 2014-05-01              6.3
#> 798 2014-06-01              6.1
#> 799 2014-07-01              6.2
#> 800 2014-08-01              6.1
#> 801 2014-09-01              5.9
#> 802 2014-10-01              5.7
#> 803 2014-11-01              5.8
#> 804 2014-12-01              5.6
#> 805 2015-01-01              5.7
#> 806 2015-02-01              5.5
#> 807 2015-03-01              5.4
#> 808 2015-04-01              5.4
#> 809 2015-05-01              5.6
#> 810 2015-06-01              5.3
#> 811 2015-07-01              5.2
#> 812 2015-08-01              5.1
#> 813 2015-09-01              5.0
#> 814 2015-10-01              5.0
#> 815 2015-11-01              5.1
#> 816 2015-12-01              5.0
#> 817 2016-01-01              4.9
#> 818 2016-02-01              4.9
#> 819 2016-03-01              5.0
#> 820 2016-04-01              5.0
#> 821 2016-05-01              4.8
#> 822 2016-06-01              4.9
#> 823 2016-07-01              4.8
#> 824 2016-08-01              4.9
#> 825 2016-09-01              5.0
#> 826 2016-10-01              4.9
#> 827 2016-11-01              4.7
#> 828 2016-12-01              4.7
#> 829 2017-01-01              4.7
#> 830 2017-02-01              4.6
#> 831 2017-03-01              4.4
#> 832 2017-04-01              4.4
#> 833 2017-05-01              4.4
#> 834 2017-06-01              4.3
#> 835 2017-07-01              4.3
#> 836 2017-08-01              4.4
#> 837 2017-09-01              4.2
#> 838 2017-10-01              4.1
#> 839 2017-11-01              4.2
#> 840 2017-12-01              4.1
#> 841 2018-01-01              4.1
#> 842 2018-02-01              4.1
#> 843 2018-03-01              4.0
#> 844 2018-04-01              4.0
#> 845 2018-05-01              3.8
#> 846 2018-06-01              4.0
#> 847 2018-07-01              3.8
#> 848 2018-08-01              3.8
#> 849 2018-09-01              3.7
#> 850 2018-10-01              3.8
#> 851 2018-11-01              3.7
#> 852 2018-12-01              3.9
#> 853 2019-01-01              4.0
#> 854 2019-02-01              3.8
#> 855 2019-03-01              3.8
#> 856 2019-04-01              3.6
#> 857 2019-05-01              3.6
#> 858 2019-06-01              3.7
#> 859 2019-07-01              3.7
#> 860 2019-08-01              3.7
#> 861 2019-09-01              3.5
#> 862 2019-10-01              3.6
#> 863 2019-11-01              3.5
#> 864 2019-12-01              3.5
#> 865 2020-01-01              3.6
#> 866 2020-02-01              3.5
#> 867 2020-03-01              4.4
#> 868 2020-04-01             14.7
#> 869 2020-05-01             13.3
#> 870 2020-06-01             11.1
#> 871 2020-07-01             10.2
#> 872 2020-08-01              8.4
#> 873 2020-09-01              7.9
#> 874 2020-10-01              6.9
#> 875 2020-11-01              6.7