Accessing FRED This function can pull time series from the FRED database: https://fred.stlouisfed.org

Accessing FRED This function can pull time series from the FRED database: https://fred.stlouisfed.org

get_fred_series2(series_id,
series_name = NULL,observation_start = NULL,
 observation_end = NULL)

Arguments

series_id
FRED times series ID.

series_name
Choose a name for the series column in output. Default: series_id.

observation_start
Choose a name for the series column in output. Default: series_id.

observation_end
Date of last observation in "yyyy-mm-dd" format. Default: Last observation available.

Examples

get_fred_series2("INDPRO", "indpro")

#>            date   indpro
#> 1    1919-01-01   5.0124
#> 2    1919-02-01   4.7908
#> 3    1919-03-01   4.6524
#> 4    1919-04-01   4.7355
#> 5    1919-05-01   4.7632
#> 6    1919-06-01   5.0678
#> 7    1919-07-01   5.3724
#> 8    1919-08-01   5.4555
#> 9    1919-09-01   5.3447
#> 10   1919-10-01   5.2893
#> 11   1919-11-01   5.2062
#> 12   1919-12-01   5.2893
#> 13   1920-01-01   5.7878
#> 14   1920-02-01   5.7878
#> 15   1920-03-01   5.6770
#> 16   1920-04-01   5.3724
#> 17   1920-05-01   5.5109
#> 18   1920-06-01   5.5663
#> 19   1920-07-01   5.4278
#> 20   1920-08-01   5.4555
#> 21   1920-09-01   5.2616
#> 22   1920-10-01   5.0401
#> 23   1920-11-01   4.6247
#> 24   1920-12-01   4.3478
#> 25   1921-01-01   4.0985
#> 26   1921-02-01   4.0155
#> 27   1921-03-01   3.9047
#> 28   1921-04-01   3.9047
#> 29   1921-05-01   4.0155
#> 30   1921-06-01   3.9878
#> 31   1921-07-01   3.9601
#> 32   1921-08-01   4.0985
#> 33   1921-09-01   4.1262
#> 34   1921-10-01   4.3755
#> 35   1921-11-01   4.3201
#> 36   1921-12-01   4.2924
#> 37   1922-01-01   4.4585
#> 38   1922-02-01   4.6524
#> 39   1922-03-01   4.9016
#> 40   1922-04-01   4.7355
#> 41   1922-05-01   4.9847
#> 42   1922-06-01   5.2339
#> 43   1922-07-01   5.2339
#> 44   1922-08-01   5.1232
#> 45   1922-09-01   5.4001
#> 46   1922-10-01   5.7047
#> 47   1922-11-01   5.9539
#> 48   1922-12-01   6.1201
#> 49   1923-01-01   5.9816
#> 50   1923-02-01   6.0647
#> 51   1923-03-01   6.2586
#> 52   1923-04-01   6.3970
#> 53   1923-05-01   6.4801
#> 54   1923-06-01   6.4247
#> 55   1923-07-01   6.3693
#> 56   1923-08-01   6.2586
#> 57   1923-09-01   6.1201
#> 58   1923-10-01   6.0924
#> 59   1923-11-01   6.0924
#> 60   1923-12-01   5.9539
#> 61   1924-01-01   6.0924
#> 62   1924-02-01   6.2032
#> 63   1924-03-01   6.0924
#> 64   1924-04-01   5.8986
#> 65   1924-05-01   5.6493
#> 66   1924-06-01   5.4001
#> 67   1924-07-01   5.3170
#> 68   1924-08-01   5.5109
#> 69   1924-09-01   5.7047
#> 70   1924-10-01   5.8432
#> 71   1924-11-01   5.9540
#> 72   1924-12-01   6.1201
#> 73   1925-01-01   6.3140
#> 74   1925-02-01   6.3140
#> 75   1925-03-01   6.3140
#> 76   1925-04-01   6.3693
#> 77   1925-05-01   6.3416
#> 78   1925-06-01   6.2863
#> 79   1925-07-01   6.4524
#> 80   1925-08-01   6.3416
#> 81   1925-09-01   6.2586
#> 82   1925-10-01   6.5078
#> 83   1925-11-01   6.6463
#> 84   1925-12-01   6.7293
#> 85   1926-01-01   6.6186
#> 86   1926-02-01   6.6186
#> 87   1926-03-01   6.7017
#> 88   1926-04-01   6.7017
#> 89   1926-05-01   6.6463
#> 90   1926-06-01   6.7293
#> 91   1926-07-01   6.7570
#> 92   1926-08-01   6.8401
#> 93   1926-09-01   6.9509
#> 94   1926-10-01   6.9509
#> 95   1926-11-01   6.9232
#> 96   1926-12-01   6.8955
#> 97   1927-01-01   6.8678
#> 98   1927-02-01   6.9232
#> 99   1927-03-01   7.0063
#> 100  1927-04-01   6.8401
#> 101  1927-05-01   6.8955
#> 102  1927-06-01   6.8678
#> 103  1927-07-01   6.7847
#> 104  1927-08-01   6.7847
#> 105  1927-09-01   6.6740
#> 106  1927-10-01   6.5355
#> 107  1927-11-01   6.5355
#> 108  1927-12-01   6.5632
#> 109  1928-01-01   6.7017
#> 110  1928-02-01   6.7570
#> 111  1928-03-01   6.8124
#> 112  1928-04-01   6.7847
#> 113  1928-05-01   6.8678
#> 114  1928-06-01   6.9232
#> 115  1928-07-01   7.0063
#> 116  1928-08-01   7.1447
#> 117  1928-09-01   7.2001
#> 118  1928-10-01   7.3386
#> 119  1928-11-01   7.4770
#> 120  1928-12-01   7.6155
#> 121  1929-01-01   7.7263
#> 122  1929-02-01   7.6986
#> 123  1929-03-01   7.7263
#> 124  1929-04-01   7.8648
#> 125  1929-05-01   8.0032
#> 126  1929-06-01   8.0586
#> 127  1929-07-01   8.1694
#> 128  1929-08-01   8.0863
#> 129  1929-09-01   8.0309
#> 130  1929-10-01   7.8924
#> 131  1929-11-01   7.5047
#> 132  1929-12-01   7.1724
#> 133  1930-01-01   7.1724
#> 134  1930-02-01   7.1447
#> 135  1930-03-01   7.0340
#> 136  1930-04-01   6.9786
#> 137  1930-05-01   6.8678
#> 138  1930-06-01   6.6740
#> 139  1930-07-01   6.3693
#> 140  1930-08-01   6.2309
#> 141  1930-09-01   6.1201
#> 142  1930-10-01   5.9539
#> 143  1930-11-01   5.8155
#> 144  1930-12-01   5.6770
#> 145  1931-01-01   5.6493
#> 146  1931-02-01   5.6770
#> 147  1931-03-01   5.7878
#> 148  1931-04-01   5.8155
#> 149  1931-05-01   5.7324
#> 150  1931-06-01   5.5939
#> 151  1931-07-01   5.5109
#> 152  1931-08-01   5.3170
#> 153  1931-09-01   5.0678
#> 154  1931-10-01   4.8739
#> 155  1931-11-01   4.8185
#> 156  1931-12-01   4.7908
#> 157  1932-01-01   4.6524
#> 158  1932-02-01   4.5416
#> 159  1932-03-01   4.4862
#> 160  1932-04-01   4.1816
#> 161  1932-05-01   4.0431
#> 162  1932-06-01   3.9047
#> 163  1932-07-01   3.7939
#> 164  1932-08-01   3.9047
#> 165  1932-09-01   4.1539
#> 166  1932-10-01   4.2924
#> 167  1932-11-01   4.2924
#> 168  1932-12-01   4.2093
#> 169  1933-01-01   4.1262
#> 170  1933-02-01   4.1539
#> 171  1933-03-01   3.9047
#> 172  1933-04-01   4.1816
#> 173  1933-05-01   4.8739
#> 174  1933-06-01   5.6216
#> 175  1933-07-01   6.1478
#> 176  1933-08-01   5.8986
#> 177  1933-09-01   5.5663
#> 178  1933-10-01   5.2893
#> 179  1933-11-01   4.9847
#> 180  1933-12-01   5.0124
#> 181  1934-01-01   5.1785
#> 182  1934-02-01   5.4278
#> 183  1934-03-01   5.6770
#> 184  1934-04-01   5.6770
#> 185  1934-05-01   5.7878
#> 186  1934-06-01   5.6770
#> 187  1934-07-01   5.2893
#> 188  1934-08-01   5.2339
#> 189  1934-09-01   4.9293
#> 190  1934-10-01   5.1509
#> 191  1934-11-01   5.2062
#> 192  1934-12-01   5.5386
#> 193  1935-01-01   5.9816
#> 194  1935-02-01   6.0924
#> 195  1935-03-01   6.0647
#> 196  1935-04-01   5.9539
#> 197  1935-05-01   5.9539
#> 198  1935-06-01   6.0370
#> 199  1935-07-01   6.0370
#> 200  1935-08-01   6.2586
#> 201  1935-09-01   6.4247
#> 202  1935-10-01   6.6186
#> 203  1935-11-01   6.7570
#> 204  1935-12-01   6.8401
#> 205  1936-01-01   6.7293
#> 206  1936-02-01   6.5632
#> 207  1936-03-01   6.6463
#> 208  1936-04-01   7.0617
#> 209  1936-05-01   7.2001
#> 210  1936-06-01   7.3386
#> 211  1936-07-01   7.4770
#> 212  1936-08-01   7.5878
#> 213  1936-09-01   7.7263
#> 214  1936-10-01   7.8371
#> 215  1936-11-01   8.0586
#> 216  1936-12-01   8.3078
#> 217  1937-01-01   8.2801
#> 218  1937-02-01   8.3909
#> 219  1937-03-01   8.5848
#> 220  1937-04-01   8.5848
#> 221  1937-05-01   8.6125
#> 222  1937-06-01   8.5017
#> 223  1937-07-01   8.5571
#> 224  1937-08-01   8.5017
#> 225  1937-09-01   8.2248
#> 226  1937-10-01   7.6155
#> 227  1937-11-01   6.8678
#> 228  1937-12-01   6.2586
#> 229  1938-01-01   6.1201
#> 230  1938-02-01   6.0647
#> 231  1938-03-01   6.0647
#> 232  1938-04-01   5.9539
#> 233  1938-05-01   5.8155
#> 234  1938-06-01   5.8709
#> 235  1938-07-01   6.2032
#> 236  1938-08-01   6.5355
#> 237  1938-09-01   6.7293
#> 238  1938-10-01   6.8955
#> 239  1938-11-01   7.1724
#> 240  1938-12-01   7.2555
#> 241  1939-01-01   7.2555
#> 242  1939-02-01   7.3109
#> 243  1939-03-01   7.3386
#> 244  1939-04-01   7.3109
#> 245  1939-05-01   7.2832
#> 246  1939-06-01   7.4494
#> 247  1939-07-01   7.6709
#> 248  1939-08-01   7.7817
#> 249  1939-09-01   8.2525
#> 250  1939-10-01   8.6678
#> 251  1939-11-01   8.8894
#> 252  1939-12-01   8.8894
#> 253  1940-01-01   8.7786
#> 254  1940-02-01   8.5017
#> 255  1940-03-01   8.3078
#> 256  1940-04-01   8.4740
#> 257  1940-05-01   8.7232
#> 258  1940-06-01   9.0002
#> 259  1940-07-01   9.1109
#> 260  1940-08-01   9.1663
#> 261  1940-09-01   9.3602
#> 262  1940-10-01   9.4986
#> 263  1940-11-01   9.7202
#> 264  1940-12-01  10.0525
#> 265  1941-01-01  10.3017
#> 266  1941-02-01  10.6063
#> 267  1941-03-01  10.9386
#> 268  1941-04-01  10.9663
#> 269  1941-05-01  11.4648
#> 270  1941-06-01  11.5479
#> 271  1941-07-01  11.6864
#> 272  1941-08-01  11.8248
#> 273  1941-09-01  11.8248
#> 274  1941-10-01  11.9356
#> 275  1941-11-01  11.9910
#> 276  1941-12-01  12.1848
#> 277  1942-01-01  12.4341
#> 278  1942-02-01  12.6556
#> 279  1942-03-01  12.7941
#> 280  1942-04-01  12.4341
#> 281  1942-05-01  12.4617
#> 282  1942-06-01  12.4895
#> 283  1942-07-01  12.7941
#> 284  1942-08-01  13.1818
#> 285  1942-09-01  13.4864
#> 286  1942-10-01  13.9295
#> 287  1942-11-01  14.2618
#> 288  1942-12-01  14.5941
#> 289  1943-01-01  14.7326
#> 290  1943-02-01  15.1203
#> 291  1943-03-01  15.2310
#> 292  1943-04-01  15.4249
#> 293  1943-05-01  15.5356
#> 294  1943-06-01  15.4526
#> 295  1943-07-01  15.9233
#> 296  1943-08-01  16.2557
#> 297  1943-09-01  16.6711
#> 298  1943-10-01  16.9203
#> 299  1943-11-01  17.1418
#> 300  1943-12-01  16.9203
#> 301  1944-01-01  17.1141
#> 302  1944-02-01  17.2526
#> 303  1944-03-01  17.2249
#> 304  1944-04-01  17.2249
#> 305  1944-05-01  17.1142
#> 306  1944-06-01  17.0588
#> 307  1944-07-01  17.0311
#> 308  1944-08-01  17.2526
#> 309  1944-09-01  17.1418
#> 310  1944-10-01  17.1972
#> 311  1944-11-01  17.0588
#> 312  1944-12-01  17.0034
#> 313  1945-01-01  16.8372
#> 314  1945-02-01  16.7818
#> 315  1945-03-01  16.6710
#> 316  1945-04-01  16.3664
#> 317  1945-05-01  15.9234
#> 318  1945-06-01  15.5633
#> 319  1945-07-01  15.2034
#> 320  1945-08-01  13.6249
#> 321  1945-09-01  12.4064
#> 322  1945-10-01  11.9079
#> 323  1945-11-01  12.3510
#> 324  1945-12-01  12.4064
#> 325  1946-01-01  11.7140
#> 326  1946-02-01  11.1325
#> 327  1946-03-01  12.2956
#> 328  1946-04-01  12.0741
#> 329  1946-05-01  11.6310
#> 330  1946-06-01  12.3510
#> 331  1946-07-01  12.7664
#> 332  1946-08-01  13.2371
#> 333  1946-09-01  13.4864
#> 334  1946-10-01  13.7356
#> 335  1946-11-01  13.8187
#> 336  1946-12-01  13.9018
#> 337  1947-01-01  14.0679
#> 338  1947-02-01  14.1510
#> 339  1947-03-01  14.2341
#> 340  1947-04-01  14.1233
#> 341  1947-05-01  14.1787
#> 342  1947-06-01  14.1787
#> 343  1947-07-01  14.0956
#> 344  1947-08-01  14.1787
#> 345  1947-09-01  14.2895
#> 346  1947-10-01  14.4279
#> 347  1947-11-01  14.6218
#> 348  1947-12-01  14.6772
#> 349  1948-01-01  14.7603
#> 350  1948-02-01  14.7880
#> 351  1948-03-01  14.6218
#> 352  1948-04-01  14.6495
#> 353  1948-05-01  14.8987
#> 354  1948-06-01  15.0926
#> 355  1948-07-01  15.0926
#> 356  1948-08-01  15.0372
#> 357  1948-09-01  14.9264
#> 358  1948-10-01  15.0372
#> 359  1948-11-01  14.8433
#> 360  1948-12-01  14.7049
#> 361  1949-01-01  14.5664
#> 362  1949-02-01  14.4280
#> 363  1949-03-01  14.1510
#> 364  1949-04-01  14.0679
#> 365  1949-05-01  13.8741
#> 366  1949-06-01  13.8464
#> 367  1949-07-01  13.8187
#> 368  1949-08-01  13.9572
#> 369  1949-09-01  14.0956
#> 370  1949-10-01  13.5695
#> 371  1949-11-01  13.9295
#> 372  1949-12-01  14.1787
#> 373  1950-01-01  14.4279
#> 374  1950-02-01  14.4833
#> 375  1950-03-01  14.9541
#> 376  1950-04-01  15.4526
#> 377  1950-05-01  15.8126
#> 378  1950-06-01  16.2834
#> 379  1950-07-01  16.8095
#> 380  1950-08-01  17.3357
#> 381  1950-09-01  17.2249
#> 382  1950-10-01  17.3357
#> 383  1950-11-01  17.3080
#> 384  1950-12-01  17.6126
#> 385  1951-01-01  17.6680
#> 386  1951-02-01  17.7788
#> 387  1951-03-01  17.8618
#> 388  1951-04-01  17.8895
#> 389  1951-05-01  17.8342
#> 390  1951-06-01  17.7511
#> 391  1951-07-01  17.4741
#> 392  1951-08-01  17.3080
#> 393  1951-09-01  17.4188
#> 394  1951-10-01  17.3911
#> 395  1951-11-01  17.5295
#> 396  1951-12-01  17.6403
#> 397  1952-01-01  17.8341
#> 398  1952-02-01  17.9449
#> 399  1952-03-01  18.0003
#> 400  1952-04-01  17.8341
#> 401  1952-05-01  17.6680
#> 402  1952-06-01  17.5018
#> 403  1952-07-01  17.2249
#> 404  1952-08-01  18.3326
#> 405  1952-09-01  18.9972
#> 406  1952-10-01  19.1911
#> 407  1952-11-01  19.5788
#> 408  1952-12-01  19.6896
#> 409  1953-01-01  19.7449
#> 410  1953-02-01  19.8557
#> 411  1953-03-01  20.0219
#> 412  1953-04-01  20.1050
#> 413  1953-05-01  20.2157
#> 414  1953-06-01  20.1326
#> 415  1953-07-01  20.3819
#> 416  1953-08-01  20.2711
#> 417  1953-09-01  19.8557
#> 418  1953-10-01  19.6896
#> 419  1953-11-01  19.2188
#> 420  1953-12-01  18.7480
#> 421  1954-01-01  18.6095
#> 422  1954-02-01  18.6649
#> 423  1954-03-01  18.5542
#> 424  1954-04-01  18.4434
#> 425  1954-05-01  18.5542
#> 426  1954-06-01  18.6096
#> 427  1954-07-01  18.6372
#> 428  1954-08-01  18.6095
#> 429  1954-09-01  18.6372
#> 430  1954-10-01  18.8588
#> 431  1954-11-01  19.1634
#> 432  1954-12-01  19.4126
#> 433  1955-01-01  19.8557
#> 434  1955-02-01  20.1050
#> 435  1955-03-01  20.5757
#> 436  1955-04-01  20.8250
#> 437  1955-05-01  21.1573
#> 438  1955-06-01  21.1850
#> 439  1955-07-01  21.3511
#> 440  1955-08-01  21.3234
#> 441  1955-09-01  21.4619
#> 442  1955-10-01  21.8219
#> 443  1955-11-01  21.8773
#> 444  1955-12-01  21.9604
#> 445  1956-01-01  22.0988
#> 446  1956-02-01  21.9050
#> 447  1956-03-01  21.9050
#> 448  1956-04-01  22.0712
#> 449  1956-05-01  21.8773
#> 450  1956-06-01  21.6835
#> 451  1956-07-01  21.0188
#> 452  1956-08-01  21.8773
#> 453  1956-09-01  22.3758
#> 454  1956-10-01  22.5696
#> 455  1956-11-01  22.3758
#> 456  1956-12-01  22.7081
#> 457  1957-01-01  22.6250
#> 458  1957-02-01  22.8466
#> 459  1957-03-01  22.8189
#> 460  1957-04-01  22.5142
#> 461  1957-05-01  22.4312
#> 462  1957-06-01  22.4866
#> 463  1957-07-01  22.6250
#> 464  1957-08-01  22.6250
#> 465  1957-09-01  22.4312
#> 466  1957-10-01  22.0988
#> 467  1957-11-01  21.5727
#> 468  1957-12-01  21.1573
#> 469  1958-01-01  20.7696
#> 470  1958-02-01  20.3265
#> 471  1958-03-01  20.0773
#> 472  1958-04-01  19.7450
#> 473  1958-05-01  19.9388
#> 474  1958-06-01  20.4650
#> 475  1958-07-01  20.7696
#> 476  1958-08-01  21.1850
#> 477  1958-09-01  21.3788
#> 478  1958-10-01  21.6281
#> 479  1958-11-01  22.2650
#> 480  1958-12-01  22.2927
#> 481  1959-01-01  22.6250
#> 482  1959-02-01  23.0681
#> 483  1959-03-01  23.4004
#> 484  1959-04-01  23.8989
#> 485  1959-05-01  24.2589
#> 486  1959-06-01  24.2866
#> 487  1959-07-01  23.7050
#> 488  1959-08-01  22.9019
#> 489  1959-09-01  22.8743
#> 490  1959-10-01  22.7081
#> 491  1959-11-01  22.8466
#> 492  1959-12-01  24.2589
#> 493  1960-01-01  24.8958
#> 494  1960-02-01  24.6743
#> 495  1960-03-01  24.4528
#> 496  1960-04-01  24.2589
#> 497  1960-05-01  24.2312
#> 498  1960-06-01  23.9266
#> 499  1960-07-01  23.8435
#> 500  1960-08-01  23.8158
#> 501  1960-09-01  23.5666
#> 502  1960-10-01  23.5389
#> 503  1960-11-01  23.2066
#> 504  1960-12-01  22.7635
#> 505  1961-01-01  22.7912
#> 506  1961-02-01  22.7635
#> 507  1961-03-01  22.9020
#> 508  1961-04-01  23.3727
#> 509  1961-05-01  23.7327
#> 510  1961-06-01  24.0650
#> 511  1961-07-01  24.3420
#> 512  1961-08-01  24.5635
#> 513  1961-09-01  24.5358
#> 514  1961-10-01  25.0066
#> 515  1961-11-01  25.3943
#> 516  1961-12-01  25.6158
#> 517  1962-01-01  25.3943
#> 518  1962-02-01  25.8097
#> 519  1962-03-01  25.9482
#> 520  1962-04-01  26.0035
#> 521  1962-05-01  25.9758
#> 522  1962-06-01  25.9205
#> 523  1962-07-01  26.1697
#> 524  1962-08-01  26.1974
#> 525  1962-09-01  26.3635
#> 526  1962-10-01  26.3912
#> 527  1962-11-01  26.5020
#> 528  1962-12-01  26.5020
#> 529  1963-01-01  26.6958
#> 530  1963-02-01  27.0005
#> 531  1963-03-01  27.1666
#> 532  1963-04-01  27.4159
#> 533  1963-05-01  27.7482
#> 534  1963-06-01  27.8313
#> 535  1963-07-01  27.7205
#> 536  1963-08-01  27.7759
#> 537  1963-09-01  28.0528
#> 538  1963-10-01  28.2466
#> 539  1963-11-01  28.3851
#> 540  1963-12-01  28.3297
#> 541  1964-01-01  28.5790
#> 542  1964-02-01  28.7728
#> 543  1964-03-01  28.7728
#> 544  1964-04-01  29.2436
#> 545  1964-05-01  29.4098
#> 546  1964-06-01  29.4928
#> 547  1964-07-01  29.6867
#> 548  1964-08-01  29.8805
#> 549  1964-09-01  29.9913
#> 550  1964-10-01  29.5759
#> 551  1964-11-01  30.4898
#> 552  1964-12-01  30.8498
#> 553  1965-01-01  31.1821
#> 554  1965-02-01  31.3760
#> 555  1965-03-01  31.7913
#> 556  1965-04-01  31.9298
#> 557  1965-05-01  32.1790
#> 558  1965-06-01  32.4283
#> 559  1965-07-01  32.7329
#> 560  1965-08-01  32.8713
#> 561  1965-09-01  32.9544
#> 562  1965-10-01  33.2868
#> 563  1965-11-01  33.4252
#> 564  1965-12-01  33.8406
#> 565  1966-01-01  34.1729
#> 566  1966-02-01  34.3945
#> 567  1966-03-01  34.8652
#> 568  1966-04-01  34.9206
#> 569  1966-05-01  35.2529
#> 570  1966-06-01  35.4191
#> 571  1966-07-01  35.6129
#> 572  1966-08-01  35.6406
#> 573  1966-09-01  35.9729
#> 574  1966-10-01  36.2221
#> 575  1966-11-01  35.9729
#> 576  1966-12-01  36.0560
#> 577  1967-01-01  36.2261
#> 578  1967-02-01  35.8152
#> 579  1967-03-01  35.6133
#> 580  1967-04-01  35.9492
#> 581  1967-05-01  35.6356
#> 582  1967-06-01  35.6311
#> 583  1967-07-01  35.5500
#> 584  1967-08-01  36.2311
#> 585  1967-09-01  36.1723
#> 586  1967-10-01  36.4669
#> 587  1967-11-01  36.9884
#> 588  1967-12-01  37.3868
#> 589  1968-01-01  37.3465
#> 590  1968-02-01  37.4804
#> 591  1968-03-01  37.5974
#> 592  1968-04-01  37.6518
#> 593  1968-05-01  38.0741
#> 594  1968-06-01  38.2140
#> 595  1968-07-01  38.1554
#> 596  1968-08-01  38.2619
#> 597  1968-09-01  38.4068
#> 598  1968-10-01  38.4830
#> 599  1968-11-01  38.9808
#> 600  1968-12-01  39.1039
#> 601  1969-01-01  39.3405
#> 602  1969-02-01  39.5922
#> 603  1969-03-01  39.9027
#> 604  1969-04-01  39.7559
#> 605  1969-05-01  39.6056
#> 606  1969-06-01  39.9925
#> 607  1969-07-01  40.2037
#> 608  1969-08-01  40.2961
#> 609  1969-09-01  40.2870
#> 610  1969-10-01  40.2988
#> 611  1969-11-01  39.9189
#> 612  1969-12-01  39.8118
#> 613  1970-01-01  39.0746
#> 614  1970-02-01  39.0488
#> 615  1970-03-01  38.9981
#> 616  1970-04-01  38.8979
#> 617  1970-05-01  38.8525
#> 618  1970-06-01  38.7269
#> 619  1970-07-01  38.8219
#> 620  1970-08-01  38.7527
#> 621  1970-09-01  38.4856
#> 622  1970-10-01  37.7157
#> 623  1970-11-01  37.4873
#> 624  1970-12-01  38.3482
#> 625  1971-01-01  38.6432
#> 626  1971-02-01  38.5695
#> 627  1971-03-01  38.5278
#> 628  1971-04-01  38.7441
#> 629  1971-05-01  38.9407
#> 630  1971-06-01  39.1040
#> 631  1971-07-01  38.9905
#> 632  1971-08-01  38.7642
#> 633  1971-09-01  39.3942
#> 634  1971-10-01  39.6891
#> 635  1971-11-01  39.8576
#> 636  1971-12-01  40.3177
#> 637  1972-01-01  41.2876
#> 638  1972-02-01  41.6865
#> 639  1972-03-01  41.9746
#> 640  1972-04-01  42.4183
#> 641  1972-05-01  42.3963
#> 642  1972-06-01  42.5166
#> 643  1972-07-01  42.4980
#> 644  1972-08-01  43.0629
#> 645  1972-09-01  43.3918
#> 646  1972-10-01  43.9718
#> 647  1972-11-01  44.4919
#> 648  1972-12-01  45.0015
#> 649  1973-01-01  45.2942
#> 650  1973-02-01  45.9656
#> 651  1973-03-01  45.9872
#> 652  1973-04-01  45.9237
#> 653  1973-05-01  46.2192
#> 654  1973-06-01  46.2509
#> 655  1973-07-01  46.4421
#> 656  1973-08-01  46.3648
#> 657  1973-09-01  46.7803
#> 658  1973-10-01  47.0954
#> 659  1973-11-01  47.3445
#> 660  1973-12-01  47.2315
#> 661  1974-01-01  46.8994
#> 662  1974-02-01  46.7538
#> 663  1974-03-01  46.7685
#> 664  1974-04-01  46.6069
#> 665  1974-05-01  46.9743
#> 666  1974-06-01  46.9233
#> 667  1974-07-01  46.9464
#> 668  1974-08-01  46.4907
#> 669  1974-09-01  46.5136
#> 670  1974-10-01  46.3401
#> 671  1974-11-01  44.8209
#> 672  1974-12-01  43.2340
#> 673  1975-01-01  42.6409
#> 674  1975-02-01  41.6602
#> 675  1975-03-01  41.2135
#> 676  1975-04-01  41.2460
#> 677  1975-05-01  41.1515
#> 678  1975-06-01  41.4224
#> 679  1975-07-01  41.8234
#> 680  1975-08-01  42.2573
#> 681  1975-09-01  42.7796
#> 682  1975-10-01  42.9687
#> 683  1975-11-01  43.0614
#> 684  1975-12-01  43.5973
#> 685  1976-01-01  44.2288
#> 686  1976-02-01  44.6725
#> 687  1976-03-01  44.7125
#> 688  1976-04-01  44.9643
#> 689  1976-05-01  45.1741
#> 690  1976-06-01  45.1830
#> 691  1976-07-01  45.4535
#> 692  1976-08-01  45.7737
#> 693  1976-09-01  45.9031
#> 694  1976-10-01  45.9207
#> 695  1976-11-01  46.5978
#> 696  1976-12-01  47.0854
#> 697  1977-01-01  46.8276
#> 698  1977-02-01  47.5417
#> 699  1977-03-01  48.1310
#> 700  1977-04-01  48.5838
#> 701  1977-05-01  48.9891
#> 702  1977-06-01  49.3428
#> 703  1977-07-01  49.4120
#> 704  1977-08-01  49.4259
#> 705  1977-09-01  49.6639
#> 706  1977-10-01  49.7592
#> 707  1977-11-01  49.8117
#> 708  1977-12-01  49.8949
#> 709  1978-01-01  49.2066
#> 710  1978-02-01  49.4502
#> 711  1978-03-01  50.3928
#> 712  1978-04-01  51.4368
#> 713  1978-05-01  51.6276
#> 714  1978-06-01  51.9833
#> 715  1978-07-01  51.9592
#> 716  1978-08-01  52.1543
#> 717  1978-09-01  52.2860
#> 718  1978-10-01  52.7054
#> 719  1978-11-01  53.1047
#> 720  1978-12-01  53.3886
#> 721  1979-01-01  53.0453
#> 722  1979-02-01  53.3303
#> 723  1979-03-01  53.4973
#> 724  1979-04-01  52.8938
#> 725  1979-05-01  53.3187
#> 726  1979-06-01  53.3157
#> 727  1979-07-01  53.2433
#> 728  1979-08-01  52.8899
#> 729  1979-09-01  52.9486
#> 730  1979-10-01  53.2390
#> 731  1979-11-01  53.1918
#> 732  1979-12-01  53.2603
#> 733  1980-01-01  53.5037
#> 734  1980-02-01  53.5053
#> 735  1980-03-01  53.3294
#> 736  1980-04-01  52.2336
#> 737  1980-05-01  50.9638
#> 738  1980-06-01  50.3348
#> 739  1980-07-01  49.9462
#> 740  1980-08-01  50.1256
#> 741  1980-09-01  50.9386
#> 742  1980-10-01  51.5813
#> 743  1980-11-01  52.4717
#> 744  1980-12-01  52.7685
#> 745  1981-01-01  52.4668
#> 746  1981-02-01  52.2260
#> 747  1981-03-01  52.5025
#> 748  1981-04-01  52.2691
#> 749  1981-05-01  52.5803
#> 750  1981-06-01  52.8284
#> 751  1981-07-01  53.1751
#> 752  1981-08-01  53.1679
#> 753  1981-09-01  52.8514
#> 754  1981-10-01  52.4949
#> 755  1981-11-01  51.8940
#> 756  1981-12-01  51.3274
#> 757  1982-01-01  50.3043
#> 758  1982-02-01  51.3016
#> 759  1982-03-01  50.9104
#> 760  1982-04-01  50.4627
#> 761  1982-05-01  50.1380
#> 762  1982-06-01  49.9692
#> 763  1982-07-01  49.8138
#> 764  1982-08-01  49.3773
#> 765  1982-09-01  49.2260
#> 766  1982-10-01  48.7874
#> 767  1982-11-01  48.5920
#> 768  1982-12-01  48.2424
#> 769  1983-01-01  49.1762
#> 770  1983-02-01  48.8688
#> 771  1983-03-01  49.2654
#> 772  1983-04-01  49.8675
#> 773  1983-05-01  50.2084
#> 774  1983-06-01  50.5089
#> 775  1983-07-01  51.2733
#> 776  1983-08-01  51.8454
#> 777  1983-09-01  52.6303
#> 778  1983-10-01  53.0679
#> 779  1983-11-01  53.2534
#> 780  1983-12-01  53.5343
#> 781  1984-01-01  54.6008
#> 782  1984-02-01  54.8350
#> 783  1984-03-01  55.1052
#> 784  1984-04-01  55.4514
#> 785  1984-05-01  55.7141
#> 786  1984-06-01  55.9085
#> 787  1984-07-01  56.0842
#> 788  1984-08-01  56.1376
#> 789  1984-09-01  56.0382
#> 790  1984-10-01  55.9459
#> 791  1984-11-01  56.1667
#> 792  1984-12-01  56.2296
#> 793  1985-01-01  56.1398
#> 794  1985-02-01  56.3323
#> 795  1985-03-01  56.4232
#> 796  1985-04-01  56.2693
#> 797  1985-05-01  56.3488
#> 798  1985-06-01  56.3901
#> 799  1985-07-01  56.0234
#> 800  1985-08-01  56.2555
#> 801  1985-09-01  56.4983
#> 802  1985-10-01  56.2648
#> 803  1985-11-01  56.4549
#> 804  1985-12-01  57.0458
#> 805  1986-01-01  57.3104
#> 806  1986-02-01  56.9344
#> 807  1986-03-01  56.5420
#> 808  1986-04-01  56.5599
#> 809  1986-05-01  56.6823
#> 810  1986-06-01  56.4976
#> 811  1986-07-01  56.8140
#> 812  1986-08-01  56.7381
#> 813  1986-09-01  56.8532
#> 814  1986-10-01  57.1191
#> 815  1986-11-01  57.3792
#> 816  1986-12-01  57.8623
#> 817  1987-01-01  57.6850
#> 818  1987-02-01  58.4399
#> 819  1987-03-01  58.5160
#> 820  1987-04-01  58.8851
#> 821  1987-05-01  59.2653
#> 822  1987-06-01  59.5409
#> 823  1987-07-01  59.9536
#> 824  1987-08-01  60.4517
#> 825  1987-09-01  60.6069
#> 826  1987-10-01  61.4910
#> 827  1987-11-01  61.8137
#> 828  1987-12-01  62.1190
#> 829  1988-01-01  62.1469
#> 830  1988-02-01  62.4169
#> 831  1988-03-01  62.5418
#> 832  1988-04-01  62.8959
#> 833  1988-05-01  62.8223
#> 834  1988-06-01  62.9823
#> 835  1988-07-01  63.0093
#> 836  1988-08-01  63.2723
#> 837  1988-09-01  63.0990
#> 838  1988-10-01  63.4122
#> 839  1988-11-01  63.5128
#> 840  1988-12-01  63.8233
#> 841  1989-01-01  64.0153
#> 842  1989-02-01  63.7242
#> 843  1989-03-01  63.8691
#> 844  1989-04-01  63.9124
#> 845  1989-05-01  63.4884
#> 846  1989-06-01  63.5187
#> 847  1989-07-01  62.9354
#> 848  1989-08-01  63.5168
#> 849  1989-09-01  63.2954
#> 850  1989-10-01  63.2542
#> 851  1989-11-01  63.4616
#> 852  1989-12-01  63.8467
#> 853  1990-01-01  63.4228
#> 854  1990-02-01  64.0446
#> 855  1990-03-01  64.3580
#> 856  1990-04-01  64.2602
#> 857  1990-05-01  64.3973
#> 858  1990-06-01  64.6041
#> 859  1990-07-01  64.5205
#> 860  1990-08-01  64.7326
#> 861  1990-09-01  64.8145
#> 862  1990-10-01  64.3274
#> 863  1990-11-01  63.5753
#> 864  1990-12-01  63.1594
#> 865  1991-01-01  62.8852
#> 866  1991-02-01  62.4462
#> 867  1991-03-01  62.1190
#> 868  1991-04-01  62.2415
#> 869  1991-05-01  62.8646
#> 870  1991-06-01  63.4372
#> 871  1991-07-01  63.5128
#> 872  1991-08-01  63.5671
#> 873  1991-09-01  64.1330
#> 874  1991-10-01  64.0213
#> 875  1991-11-01  63.9480
#> 876  1991-12-01  63.6937
#> 877  1992-01-01  63.3374
#> 878  1992-02-01  63.7911
#> 879  1992-03-01  64.3212
#> 880  1992-04-01  64.8088
#> 881  1992-05-01  65.0202
#> 882  1992-06-01  65.0295
#> 883  1992-07-01  65.6172
#> 884  1992-08-01  65.2943
#> 885  1992-09-01  65.4460
#> 886  1992-10-01  65.9369
#> 887  1992-11-01  66.2174
#> 888  1992-12-01  66.2772
#> 889  1993-01-01  66.5643
#> 890  1993-02-01  66.8594
#> 891  1993-03-01  66.7658
#> 892  1993-04-01  67.0000
#> 893  1993-05-01  66.7674
#> 894  1993-06-01  66.8781
#> 895  1993-07-01  67.0845
#> 896  1993-08-01  67.0140
#> 897  1993-09-01  67.3345
#> 898  1993-10-01  67.8515
#> 899  1993-11-01  68.1326
#> 900  1993-12-01  68.5054
#> 901  1994-01-01  68.7648
#> 902  1994-02-01  68.7836
#> 903  1994-03-01  69.4766
#> 904  1994-04-01  69.8703
#> 905  1994-05-01  70.2281
#> 906  1994-06-01  70.6763
#> 907  1994-07-01  70.7846
#> 908  1994-08-01  71.2053
#> 909  1994-09-01  71.4767
#> 910  1994-10-01  72.0820
#> 911  1994-11-01  72.5263
#> 912  1994-12-01  73.2879
#> 913  1995-01-01  73.4219
#> 914  1995-02-01  73.3025
#> 915  1995-03-01  73.4081
#> 916  1995-04-01  73.3612
#> 917  1995-05-01  73.6104
#> 918  1995-06-01  73.8602
#> 919  1995-07-01  73.5664
#> 920  1995-08-01  74.4950
#> 921  1995-09-01  74.7937
#> 922  1995-10-01  74.7017
#> 923  1995-11-01  74.8897
#> 924  1995-12-01  75.1755
#> 925  1996-01-01  74.6841
#> 926  1996-02-01  75.8344
#> 927  1996-03-01  75.7631
#> 928  1996-04-01  76.4562
#> 929  1996-05-01  77.0161
#> 930  1996-06-01  77.6669
#> 931  1996-07-01  77.5662
#> 932  1996-08-01  78.0160
#> 933  1996-09-01  78.5532
#> 934  1996-10-01  78.5065
#> 935  1996-11-01  79.1996
#> 936  1996-12-01  79.7143
#> 937  1997-01-01  79.8273
#> 938  1997-02-01  80.7930
#> 939  1997-03-01  81.3340
#> 940  1997-04-01  81.3531
#> 941  1997-05-01  81.8293
#> 942  1997-06-01  82.2285
#> 943  1997-07-01  82.8557
#> 944  1997-08-01  83.7214
#> 945  1997-09-01  84.4651
#> 946  1997-10-01  85.1918
#> 947  1997-11-01  85.9397
#> 948  1997-12-01  86.2047
#> 949  1998-01-01  86.6474
#> 950  1998-02-01  86.7612
#> 951  1998-03-01  86.8198
#> 952  1998-04-01  87.1411
#> 953  1998-05-01  87.6952
#> 954  1998-06-01  87.1450
#> 955  1998-07-01  86.8423
#> 956  1998-08-01  88.6247
#> 957  1998-09-01  88.4515
#> 958  1998-10-01  89.1677
#> 959  1998-11-01  89.1098
#> 960  1998-12-01  89.4407
#> 961  1999-01-01  89.8594
#> 962  1999-02-01  90.3386
#> 963  1999-03-01  90.4819
#> 964  1999-04-01  90.7274
#> 965  1999-05-01  91.3520
#> 966  1999-06-01  91.1994
#> 967  1999-07-01  91.7766
#> 968  1999-08-01  92.1629
#> 969  1999-09-01  91.7740
#> 970  1999-10-01  92.9979
#> 971  1999-11-01  93.4371
#> 972  1999-12-01  94.1593
#> 973  2000-01-01  94.1758
#> 974  2000-02-01  94.4557
#> 975  2000-03-01  94.7980
#> 976  2000-04-01  95.4808
#> 977  2000-05-01  95.6435
#> 978  2000-06-01  95.7353
#> 979  2000-07-01  95.5906
#> 980  2000-08-01  95.3112
#> 981  2000-09-01  95.6790
#> 982  2000-10-01  95.3970
#> 983  2000-11-01  95.4229
#> 984  2000-12-01  95.1573
#> 985  2001-01-01  94.5448
#> 986  2001-02-01  93.9398
#> 987  2001-03-01  93.7201
#> 988  2001-04-01  93.4469
#> 989  2001-05-01  92.8760
#> 990  2001-06-01  92.3208
#> 991  2001-07-01  91.7933
#> 992  2001-08-01  91.6795
#> 993  2001-09-01  91.3289
#> 994  2001-10-01  90.9315
#> 995  2001-11-01  90.4860
#> 996  2001-12-01  90.5073
#> 997  2002-01-01  91.0794
#> 998  2002-02-01  91.0553
#> 999  2002-03-01  91.7980
#> 1000 2002-04-01  92.1771
#> 1001 2002-05-01  92.5668
#> 1002 2002-06-01  93.4476
#> 1003 2002-07-01  93.2237
#> 1004 2002-08-01  93.2359
#> 1005 2002-09-01  93.3654
#> 1006 2002-10-01  93.0834
#> 1007 2002-11-01  93.5693
#> 1008 2002-12-01  93.1103
#> 1009 2003-01-01  93.8198
#> 1010 2003-02-01  93.9532
#> 1011 2003-03-01  93.7358
#> 1012 2003-04-01  93.0657
#> 1013 2003-05-01  93.0918
#> 1014 2003-06-01  93.2476
#> 1015 2003-07-01  93.6582
#> 1016 2003-08-01  93.5246
#> 1017 2003-09-01  94.0751
#> 1018 2003-10-01  94.2079
#> 1019 2003-11-01  94.9338
#> 1020 2003-12-01  94.8662
#> 1021 2004-01-01  95.1085
#> 1022 2004-02-01  95.6847
#> 1023 2004-03-01  95.2088
#> 1024 2004-04-01  95.6385
#> 1025 2004-05-01  96.3966
#> 1026 2004-06-01  95.6019
#> 1027 2004-07-01  96.3385
#> 1028 2004-08-01  96.4095
#> 1029 2004-09-01  96.4907
#> 1030 2004-10-01  97.4094
#> 1031 2004-11-01  97.6137
#> 1032 2004-12-01  98.3318
#> 1033 2005-01-01  98.7845
#> 1034 2005-02-01  99.4639
#> 1035 2005-03-01  99.3061
#> 1036 2005-04-01  99.4734
#> 1037 2005-05-01  99.6033
#> 1038 2005-06-01  99.9853
#> 1039 2005-07-01  99.6692
#> 1040 2005-08-01  99.9435
#> 1041 2005-09-01  98.0779
#> 1042 2005-10-01  99.3149
#> 1043 2005-11-01 100.3216
#> 1044 2005-12-01 100.9437
#> 1045 2006-01-01 101.0627
#> 1046 2006-02-01 101.0669
#> 1047 2006-03-01 101.2751
#> 1048 2006-04-01 101.6813
#> 1049 2006-05-01 101.5790
#> 1050 2006-06-01 101.9693
#> 1051 2006-07-01 101.9319
#> 1052 2006-08-01 102.3327
#> 1053 2006-09-01 102.1252
#> 1054 2006-10-01 102.0668
#> 1055 2006-11-01 101.9688
#> 1056 2006-12-01 103.0292
#> 1057 2007-01-01 102.4933
#> 1058 2007-02-01 103.5264
#> 1059 2007-03-01 103.7521
#> 1060 2007-04-01 104.4814
#> 1061 2007-05-01 104.5322
#> 1062 2007-06-01 104.5617
#> 1063 2007-07-01 104.5227
#> 1064 2007-08-01 104.7556
#> 1065 2007-09-01 105.1611
#> 1066 2007-10-01 104.7171
#> 1067 2007-11-01 105.3338
#> 1068 2007-12-01 105.3457
#> 1069 2008-01-01 105.0619
#> 1070 2008-02-01 104.7094
#> 1071 2008-03-01 104.4616
#> 1072 2008-04-01 103.6704
#> 1073 2008-05-01 103.0863
#> 1074 2008-06-01 102.8445
#> 1075 2008-07-01 102.3002
#> 1076 2008-08-01 100.7353
#> 1077 2008-09-01  96.3666
#> 1078 2008-10-01  97.2832
#> 1079 2008-11-01  96.0605
#> 1080 2008-12-01  93.2521
#> 1081 2009-01-01  91.0373
#> 1082 2009-02-01  90.4502
#> 1083 2009-03-01  89.0160
#> 1084 2009-04-01  88.3056
#> 1085 2009-05-01  87.4155
#> 1086 2009-06-01  87.0742
#> 1087 2009-07-01  88.0323
#> 1088 2009-08-01  89.0190
#> 1089 2009-09-01  89.6926
#> 1090 2009-10-01  89.9797
#> 1091 2009-11-01  90.3375
#> 1092 2009-12-01  90.6132
#> 1093 2010-01-01  91.6642
#> 1094 2010-02-01  91.9940
#> 1095 2010-03-01  92.5993
#> 1096 2010-04-01  92.9436
#> 1097 2010-05-01  94.2997
#> 1098 2010-06-01  94.4397
#> 1099 2010-07-01  94.8536
#> 1100 2010-08-01  95.1448
#> 1101 2010-09-01  95.3637
#> 1102 2010-10-01  95.1109
#> 1103 2010-11-01  95.1383
#> 1104 2010-12-01  96.0599
#> 1105 2011-01-01  95.9364
#> 1106 2011-02-01  95.5154
#> 1107 2011-03-01  96.4643
#> 1108 2011-04-01  96.1187
#> 1109 2011-05-01  96.3377
#> 1110 2011-06-01  96.6154
#> 1111 2011-07-01  97.1292
#> 1112 2011-08-01  97.6731
#> 1113 2011-09-01  97.6494
#> 1114 2011-10-01  98.3222
#> 1115 2011-11-01  98.2433
#> 1116 2011-12-01  98.7876
#> 1117 2012-01-01  99.3925
#> 1118 2012-02-01  99.6203
#> 1119 2012-03-01  99.1551
#> 1120 2012-04-01  99.9006
#> 1121 2012-05-01 100.0924
#> 1122 2012-06-01 100.0728
#> 1123 2012-07-01 100.3354
#> 1124 2012-08-01  99.8560
#> 1125 2012-09-01  99.9049
#> 1126 2012-10-01 100.1167
#> 1127 2012-11-01 100.5991
#> 1128 2012-12-01 100.9542
#> 1129 2013-01-01 100.8204
#> 1130 2013-02-01 101.3995
#> 1131 2013-03-01 101.8114
#> 1132 2013-04-01 101.6364
#> 1133 2013-05-01 101.7476
#> 1134 2013-06-01 101.9549
#> 1135 2013-07-01 101.5204
#> 1136 2013-08-01 102.1907
#> 1137 2013-09-01 102.7170
#> 1138 2013-10-01 102.5371
#> 1139 2013-11-01 102.8379
#> 1140 2013-12-01 103.1504
#> 1141 2014-01-01 102.7216
#> 1142 2014-02-01 103.5917
#> 1143 2014-03-01 104.5889
#> 1144 2014-04-01 104.6371
#> 1145 2014-05-01 105.0152
#> 1146 2014-06-01 105.4081
#> 1147 2014-07-01 105.6157
#> 1148 2014-08-01 105.4986
#> 1149 2014-09-01 105.8138
#> 1150 2014-10-01 105.8357
#> 1151 2014-11-01 106.6634
#> 1152 2014-12-01 106.5085
#> 1153 2015-01-01 105.9806
#> 1154 2015-02-01 105.4425
#> 1155 2015-03-01 105.1464
#> 1156 2015-04-01 104.5272
#> 1157 2015-05-01 104.0742
#> 1158 2015-06-01 103.7174
#> 1159 2015-07-01 104.3243
#> 1160 2015-08-01 104.1621
#> 1161 2015-09-01 103.7768
#> 1162 2015-10-01 103.3976
#> 1163 2015-11-01 102.6866
#> 1164 2015-12-01 102.1014
#> 1165 2016-01-01 102.9525
#> 1166 2016-02-01 102.2225
#> 1167 2016-03-01 101.4155
#> 1168 2016-04-01 101.5167
#> 1169 2016-05-01 101.4298
#> 1170 2016-06-01 101.8747
#> 1171 2016-07-01 102.1325
#> 1172 2016-08-01 102.0407
#> 1173 2016-09-01 102.0477
#> 1174 2016-10-01 102.2485
#> 1175 2016-11-01 102.0507
#> 1176 2016-12-01 102.9281
#> 1177 2017-01-01 103.0366
#> 1178 2017-02-01 102.6479
#> 1179 2017-03-01 103.3430
#> 1180 2017-04-01 104.2721
#> 1181 2017-05-01 104.4129
#> 1182 2017-06-01 104.5849
#> 1183 2017-07-01 104.5427
#> 1184 2017-08-01 104.0475
#> 1185 2017-09-01 104.0502
#> 1186 2017-10-01 105.6287
#> 1187 2017-11-01 106.1930
#> 1188 2017-12-01 106.5360
#> 1189 2018-01-01 106.2655
#> 1190 2018-02-01 106.6419
#> 1191 2018-03-01 107.2519
#> 1192 2018-04-01 108.2223
#> 1193 2018-05-01 107.3639
#> 1194 2018-06-01 108.1707
#> 1195 2018-07-01 108.6520
#> 1196 2018-08-01 109.5246
#> 1197 2018-09-01 109.6749
#> 1198 2018-10-01 109.9165
#> 1199 2018-11-01 110.5067
#> 1200 2018-12-01 110.5516
#> 1201 2019-01-01 110.1185
#> 1202 2019-02-01 109.5631
#> 1203 2019-03-01 109.6811
#> 1204 2019-04-01 108.9888
#> 1205 2019-05-01 109.2264
#> 1206 2019-06-01 109.2774
#> 1207 2019-07-01 109.0852
#> 1208 2019-08-01 109.9634
#> 1209 2019-09-01 109.4437
#> 1210 2019-10-01 108.8532
#> 1211 2019-11-01 109.7573
#> 1212 2019-12-01 109.4330