Search series_id by key words

Get US federal reserve economic data series that match keywords This wrapper function is to access data from https://alfred.stlouisfed.org

search_fedeconseries(search_text, search_type = 'full_text',
                            realtime_start = NULL, realtime_end = NULL)

Arguments

search_text
The words to match economic data series; key words.

search_type
Determines the type of search to perform. One of the following strings: 'full_text', 'series_id'.

realtime_start
Date of first real time period in "yyyy-mm-dd" format, default: today's date.

realtime_end
Date of last real time period in "yyyy-mm-dd" format, default: today's date.

Value

return dataframe of variables ids.

Note

More arguments could be added into function, for additional arguments, check https://research.stlouisfed.org/docs/api/fred/series_search.html

Examples

search_fedeconseries('GDP', search_type = 'full_text',realtime_start = NULL, realtime_end = NULL)

#> Error in open.connection(con, "rb") : HTTP error 504.

#> Error in search_fedeconseries("GDP", search_type = "full_text", realtime_start = NULL,     realtime_end = NULL): Fail to fetch search results, please check if parameters are right?